renamed SearchBar.js to index.js and fixed applicable imports
