+++ client/src/components/SearchBar/index.js
